Basic program to play Tic Tac Toe with computer.
This is a basic algorithm to play Tic Tac Toe with the computer. By default the computer uses Os and the user uses Xs. The responses of the computer aren't based on machine learning , they are solely based on some random set of coordinates. I need to improve the algorithm with some machine learning.